Updates action not to place retype.json at temp folder.
The script will no longer switch to temp folder when running
retype-init, and also changes actual output path to the temp folder's
root, then checking, instead of directory existence, for config.js after
retype is built.

Previously, the <temp>/output folder was created by retype. To avoid
temporary folder conflicts, it is created as soon as the temporary name
is generated, so it will always be present after a retype-build
completes.

This change is necessary because it no longer makes sense to change
retype.json now that the --override commandline argument is supported by
Retype.

Related GitHub issue: #1.