Creates publish action and script.
The action will fetch latest Retype version off NuGet website and patch
the build.sh script to point to it.

Then it will create or update the git tags as needed.

Finally, it will either create a release or update an existing one
(which would be turned into draft because its tag gets temporarily
deleted from GitHub), ensuring the 'v<major version number>' tag is
up-to-date with the latest changes.

The action is set up to be triggered via workflow dispatch event, i.e.
is run manually or from external program.