Pulls the respective platform's NPM package if no dotnet.
Currently the script will pull the multi-platform NPM package in case
it cannot find the right version of .NET installed in the system.

This changes the build script so that it pulls one of the platform
specific package (retype-win-x64, retype-linux-x64, retype-darwin-x64)
and use it, significantly reducing the download size.

Related GitHub issue: retypeapp/action-build#14.