Properly checks whether no config file is found in repo.
Now checks correctly whether the output of the find command is empty to
assume no config could be found and initialize with default one.

The script was returning an empty result which counted as "one line
result", then an empty path was being assigned as config file in case no
valid one was found, thus passing an empty parameter to retype startup
leading to failure.

Related GitHub issue: retypeapp/retype#257.