Fixes issues with retype.yml config file locate algorithm.
- Hidden paths (starting with a dot) will no longer be used while
  searching for Retype configuration files.

This will avoid, for instance, selecting workflow files that matches the
config file name when no config file is on root of the repo.

- Properly checks whether no file is found in case there's no file in
  root and repository search was performed.

An empty file list result would also count as "1 line" result, thus
checking for an empty result is necessary, instead of "zero lines" in
the output variable.

Related GitHub issue: retypeapp/retype#257.